feat(Collection): add basic story
