+++ packages/components/collection/src/stories/CollectionDemo.vue
@@ -1,7 +1,9 @@
 <!-- eslint-disable no-console -->
 <script setup lang="ts">
 import { onMounted, ref } from 'vue'
-import { createCollection } from '@oku-ui/collection'
+import List from './List.vue'
+import Item from './Item.vue'
+import { LogItems } from './LogItems'
 
 export interface OkuCollectionProps {
   template: '#1' | '#2'
@@ -12,48 +14,23 @@ withDefaults(defineProps<OkuCollectionProps>(), {
   template: '#1',
 })
 
-type ItemData = { disabled?: boolean }
-
-const { CollectionSlot, CollectionItemSlot, CollectionProvider, useCollection } = createCollection<HTMLLIElement, ItemData >('List')
-
 const labelRef = ref<any>()
 onMounted(() => {
   console.log(labelRef.value, 'ref')
 })
-const alert = () => window.alert('clicked')
-
-function LogsItem() {
-  const getItems = useCollection(undefined)
-  console.log(getItems.value)
-}
 </script>
 
 <template>
   <div class="cursor-default inline-block">
     <div v-if="template === '#1' || allshow" class="flex flex-col">
-      <CollectionProvider :scope="undefined">
-        <CollectionSlot :scope="undefined">
-          <ul clas="w-52">
-            <CollectionItemSlot ref="labelRef" :scope="undefined">
-              <li>
-                Red
-              </li>
-            </CollectionItemSlot>
-            <CollectionItemSlot :scope="undefined" :disabled="true">
-              <li class="opacity-50">
-                Green
-              </li>
-            </CollectionItemSlot>
-
-            <CollectionItemSlot :scope="undefined">
-              <li>
-                Blue
-              </li>
-            </CollectionItemSlot>
-          </ul>
-        </CollectionSlot>
-        <LogsItem />
-      </CollectionProvider>
+      <List>
+        <Item>Red</Item>
+        <Item disabled>
+          Green
+        </Item>
+        <Item>Blue</Item>
+        <LogItems />
+      </List>
     </div>
   </div>
 </template>
