+++ packages/components/collection/src/stories/List.vue
@@ -1,0 +1,13 @@
+<script lang="ts" setup>
+import { CollectionProvider, CollectionSlot } from './utils'
+</script>
+
+<template>
+  <CollectionProvider :scope="undefined">
+    <CollectionSlot :scope="undefined">
+      <ul style="width: 200px;">
+        <slot />
+      </ul>
+    </CollectionSlot>
+  </CollectionProvider>
+</template>
