+++ packages/components/collection/src/stories/utils.ts
@@ -1,0 +1,5 @@
+import { createCollection } from '../collection'
+
+type ItemData = { disabled: boolean }
+
+export const { CollectionProvider, CollectionSlot, CollectionItemSlot, useCollection } = createCollection<HTMLElement, ItemData>('List')
